+++ GloryEngine/Editor/GloryEditor/ObjectMenu.cpp
@@ -72,7 +72,7 @@ namespace Glory::Editor
 		std::vector<std::string> result;
 		std::string currentPath = path;
 
-		int forwardSlashIndex = currentPath.find("/");
+		size_t forwardSlashIndex = currentPath.find("/");
 		while (forwardSlashIndex != std::string::npos)
 		{
 			std::string item = currentPath.substr(0, forwardSlashIndex);
