+++ GloryEngine/Engine/Modules/GloryEntityScenes/ScriptedComponentSerializer.cpp
@@ -6,7 +6,7 @@ namespace Glory
 	void ScriptedComponentSerailizer::Serialize(UUID uuid, ScriptedComponent& component, YAML::Emitter& out)
 	{
 		const GloryReflect::TypeData* pTypeData = ScriptedComponent::GetTypeData();
-		for (uint32_t i = 0; i < pTypeData->FieldCount(); i++)
+		for (size_t i = 0; i < pTypeData->FieldCount(); i++)
 		{
 			const GloryReflect::FieldData* pFieldData = pTypeData->GetFieldData(i);
 			PropertySerializer::SerializeProperty(pFieldData, &component, out);
@@ -18,7 +18,7 @@ namespace Glory
 	void ScriptedComponentSerailizer::Deserialize(ScriptedComponent& component, YAML::Node& object, Serializer::Flags flags)
 	{
 		const GloryReflect::TypeData* pTypeData = ScriptedComponent::GetTypeData();
-		for (uint32_t i = 0; i < pTypeData->FieldCount(); i++)
+		for (size_t i = 0; i < pTypeData->FieldCount(); i++)
 		{
 			const GloryReflect::FieldData* pFieldData = pTypeData->GetFieldData(i);
 			PropertySerializer::DeserializeProperty(pFieldData, &component, object[pFieldData->Name()]);
