+++ GloryEngine/Engine/Modules/GloryOpenGLGraphics/GLMesh.cpp
@@ -5,7 +5,7 @@
 
 namespace Glory
 {
-	GLMesh::GLMesh(size_t vertexCount, size_t indexCount, InputRate inputRate, size_t binding, size_t stride, const std::vector<AttributeType>& attributeTypes) :
+	GLMesh::GLMesh(uint32_t vertexCount, uint32_t indexCount, InputRate inputRate, uint32_t binding, uint32_t stride, const std::vector<AttributeType>& attributeTypes) :
 		Mesh(vertexCount, indexCount, inputRate, binding, stride, attributeTypes), m_VertexArrayID(0)
 	{
 	}
